+++ SDK-sources/OutbrainSDK/Viewability/OBViewabilityService.m
@@ -103,6 +103,11 @@ float const kThirtyMinutesInSeconds = 30.0 * 60.0;
     NSTimeInterval timeIntervalSinceRequestStart = [timeNow timeIntervalSinceDate:requestStartDate];
     NSString *timeToProcessRequest = [NSString stringWithFormat:@"%d", (long) (timeIntervalSinceRequestStart * 1000)];
     
+    if (viewabilityData.reportServedUrl == nil) {
+        NSLog(@"Error - reportRecsReceived, reportServedUrl is nil");
+        return;
+    }
+    
     NSString *viewabilityUrl = [self editTmParameterInUrl:viewabilityData.reportServedUrl tm:timeToProcessRequest];
     
     NSURLComponents *components = [NSURLComponents componentsWithString:viewabilityUrl];
@@ -168,6 +173,11 @@ float const kThirtyMinutesInSeconds = 30.0 * 60.0;
         
         [self.reqIdAlreadyReportedArray addObject:viewabilityData.rId];
         
+        if (viewabilityData.reportViewedUrl == nil) {
+            NSLog(@"Error - reportRecsShownForKey, reportViewedUrl is nil");
+            return;
+        }
+        
         NSString *viewabilityUrl = [self editTmParameterInUrl:viewabilityData.reportViewedUrl tm:executionTime];
         
         NSURLComponents *components = [NSURLComponents componentsWithString:viewabilityUrl];
@@ -204,12 +214,20 @@ float const kThirtyMinutesInSeconds = 30.0 * 60.0;
 
 
 #pragma mark - Private
--(NSString *) editTmParameterInUrl:(NSString *)url tm:(NSString *)tm {
-    NSString *tmString = [@"tm=" stringByAppendingString:tm];
-    if ([url containsString:@"tm=0"]) {
-        return [url stringByReplacingOccurrencesOfString:@"tm=0" withString:tmString];
+-(NSString *) editTmParameterInUrl:(NSString *)urlString tm:(NSString *)tm {
+    NSString *tmQueryString = [@"tm=" stringByAppendingString:tm];
+    if ([urlString containsString:@"tm=0"]) {
+        return [urlString stringByReplacingOccurrencesOfString:@"tm=0" withString:tmQueryString];
     } else {
-        return tmString;
+        NSURL *url = [NSURL URLWithString:urlString];
+        
+        if (![urlString length]) {
+            return urlString;
+        }
+        
+        NSString *URLString = [[NSString alloc] initWithFormat:@"%@%@%@", [url absoluteString],
+                               url.query ? @"&" : @"?", tmQueryString];
+        return [[NSURL URLWithString:URLString] absoluteString];
     }
 }